Change telegram_id column type to BigInteger

+++ backend/alembic/versions/001_initial_schema.py
@@ -21,7 +21,7 @@ def upgrade() -> None:
     op.create_table(
         "users",
         sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("telegram_id", sa.Integer(), nullable=False),
+        sa.Column("telegram_id", sa.BigInteger(), nullable=False),
         sa.Column("first_name", sa.String(length=128), nullable=True),
         sa.Column("last_name", sa.String(length=128), nullable=True),
         sa.Column("username", sa.String(length=64), nullable=True),
